+++ example/01-bootstrap.php
@@ -12,8 +12,8 @@ use PhpAb\Storage\Adapter\Cookie;
 use PhpAb\Storage\Storage;
 use PhpAb\Participation\Manager;
 use PhpAb\Analytics\DataCollector\Google;
-use PhpAb\Participation\Filter\Percentage;
-use PhpAb\Variant\Chooser\RandomChooser;
+use PhpAb\Filter\Percentage;
+use PhpAb\Chooser\RandomChooser;
 use PhpAb\Engine\Engine;
 use PhpAb\Test\Test;
 use PhpAb\Variant\SimpleVariant;
